requirements.txt
numpy==2.1.0
pyyaml==6.0.2